+++ baybe/transformations/composite.py
@@ -1,17 +1,21 @@
 """Composite transformations."""
 
+from __future__ import annotations
+
 from functools import reduce
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from attrs import define, field
 from attrs.validators import and_, deep_iterable, instance_of, max_len, min_len
-from torch import Tensor
 from typing_extensions import override
 
 from baybe.transformations.base import Transformation
 from baybe.transformations.utils import compress_transformations
-from baybe.utils.basic import compose
+from baybe.utils.basic import compose, to_tuple
 from baybe.utils.interval import Interval
+
+if TYPE_CHECKING:
+    from torch import Tensor
 
 
 @define
@@ -29,15 +33,19 @@ class ChainedTransformation(Transformation):
 
     @override
     def __eq__(self, other: Any, /) -> bool:
-        # A chained transformation with only one element is equivalent to that element
         if len(self.transformations) == 1:
+            # A chained transformation with only one element is equivalent to that
+            # element
             return self.transformations[0] == other
+        if isinstance(other, ChainedTransformation):
+            return self.transformations == other.transformations
+        return NotImplemented
 
-        # TODO: https://github.com/python-attrs/attrs/issues/1452
-        return (
-            super().__eq__(other)
-            and isinstance(other, ChainedTransformation)
-            and self.transformations == other.transformations
+    @override
+    def get_codomain(self, interval: Interval | None = None, /) -> Interval:
+        interval = Interval.create(interval)
+        return reduce(
+            lambda acc, t: t.get_codomain(acc), self.transformations, interval
         )
 
     @override
@@ -55,7 +63,7 @@ class AdditiveTransformation(Transformation):
     """A transformation implementing the sum of two transformations."""
 
     transformations: tuple[Transformation, Transformation] = field(
-        converter=tuple,
+        converter=to_tuple,
         validator=deep_iterable(
             iterable_validator=and_(min_len(2), max_len(2)),
             member_validator=instance_of(Transformation),
@@ -64,17 +72,11 @@ class AdditiveTransformation(Transformation):
     """The transformations to be added."""
 
     @override
-    def get_image(self, interval: Interval | None = None, /) -> Interval:
-        # NOTE: This only provides a conservative estimate of the image in that it
-        #   computes an "upper bound" (i.e. an interval that contains the actual image),
-        #   which is produced under the assumption that the extremal values of the two
-        #   individual transformations occur at the same input value. Computing the
-        #   exact image without additional information would require evaluating the
-        #   transformation on the entire (uncountable) input space, which is infeasible.
+    def get_codomain(self, interval: Interval | None = None, /) -> Interval:
         interval = Interval.create(interval)
-        im1 = self.transformations[0].get_image(interval)
-        im2 = self.transformations[1].get_image(interval)
-        return Interval([im1.lower + im2.lower, im1.upper + im2.upper])
+        im1 = self.transformations[0].get_codomain(interval)
+        im2 = self.transformations[1].get_codomain(interval)
+        return Interval(im1.lower + im2.lower, im1.upper + im2.upper)
 
     @override
     def __call__(self, x: Tensor, /) -> Tensor:
@@ -86,7 +88,7 @@ class MultiplicativeTransformation(Transformation):
     """A transformation implementing the product of two transformations."""
 
     transformations: tuple[Transformation, Transformation] = field(
-        converter=tuple,
+        converter=to_tuple,
         validator=deep_iterable(
             iterable_validator=and_(min_len(2), max_len(2)),
             member_validator=instance_of(Transformation),
@@ -95,16 +97,10 @@ class MultiplicativeTransformation(Transformation):
     """The transformations to be multiplied."""
 
     @override
-    def get_image(self, interval: Interval | None = None, /) -> Interval:
-        # NOTE: This only provides a conservative estimate of the image in that it
-        #   computes an "upper bound" (i.e. an interval that contains the actual image),
-        #   which is produced under the assumption that the extremal values of the two
-        #   individual transformations occur at the same input value. Computing the
-        #   exact image without additional information would require evaluating the
-        #   transformation on the entire (uncountable) input space, which is infeasible.
+    def get_codomain(self, interval: Interval | None = None, /) -> Interval:
         interval = Interval.create(interval)
-        im1 = self.transformations[0].get_image(interval)
-        im2 = self.transformations[1].get_image(interval)
+        im1 = self.transformations[0].get_codomain(interval)
+        im2 = self.transformations[1].get_codomain(interval)
         boundary_products = [
             im1.lower * im2.lower,
             im1.lower * im2.upper,