+++ tests/insights/test_shap.py
@@ -118,8 +118,15 @@ def _test_shap_insight(campaign, explainer_cls, use_comp_rep, is_shap):
         if use_comp_rep:
             df = campaign.searchspace.transform(df)
         shap_explanations = shap_insight.explain(df)
+        assert len(shap_explanations) == campaign.objective._n_models
         assert isinstance(shap_explanations, tuple)
         assert is_all_instance(shap_explanations, shap.Explanation)
+
+
+_desirabilty_targets = [
+    NumericalTarget.normalize_ramp("t1", cutoffs=(0, 100)),
+    NumericalTarget.normalize_ramp("t2", cutoffs=(0, 100), descending=True),
+]
 
 
 @mark.slow
@@ -128,15 +135,12 @@ def _test_shap_insight(campaign, explainer_cls, use_comp_rep, is_shap):
     [
         param(SingleTargetObjective(NumericalTarget("t1")), id="single"),
         param(
-            DesirabilityObjective(
-                (
-                    NumericalTarget.normalize_ramp("t1", cutoffs=(0, 100)),
-                    NumericalTarget.normalize_ramp(
-                        "t2", cutoffs=(0, 100), descending=True
-                    ),
-                )
-            ),
-            id="desirability",
+            DesirabilityObjective(_desirabilty_targets, as_pre_transformation=False),
+            id="desirability_post",
+        ),
+        param(
+            DesirabilityObjective(_desirabilty_targets, as_pre_transformation=True),
+            id="desirability_pre",
         ),
         param(
             ParetoObjective(